+++ src/App.tsx
@@ -3,8 +3,6 @@ import { Viewer } from './lib/components/Viewer'
 import { pdf_data } from './examples/pdf/latexSamplePDF'
 import { exampleShareFunction } from './examples/function/sharedFunction'
 import { exampleCitation } from './examples/citation/citation'
-import Introduction from './lib/components/helpers/introduction/Introduction'
-import { useState } from 'react'
 
 const homeFunction = () => {
   console.log('home')